Second Module creation
